requirements.txt
Flask==2.3.2
flask-swagger-ui==4.11.1
Pillow
ImageHash
